Rename and relocate MenuGroup to PopoverGroup
